mas cambios

+++ src/compras/CompraPuntual.java
@@ -3,18 +3,18 @@ package compras;
 import puntoDeVenta.PuntoDeVenta;
 
 public class CompraPuntual extends Compra {
-	private Double cantidadDeHoras;
+	private Integer cantidadDeHoras;
 
-	public CompraPuntual(PuntoDeVenta puntoDeVenta, Double cantidadDeHoras) {
+	public CompraPuntual(PuntoDeVenta puntoDeVenta, Integer cantidadDeHoras) {
 		super(puntoDeVenta);
 		this.setCantidadDeHoras(cantidadDeHoras);
 	}
 
-	public Double getCantidadDeHoras() {
+	public Integer getCantidadDeHoras() {
 		return cantidadDeHoras;
 	}
 
-	private void setCantidadDeHoras(Double cantidadDeHoras) {
+	private void setCantidadDeHoras(Integer cantidadDeHoras) {
 		this.cantidadDeHoras = cantidadDeHoras;
 	}
 }
